requirements.txt
#Framework Web
flask==3.0.0
flask-cors==6.0.2
werkzeug==3.1.5

tensorflow==2.20.0
tensorflow_keras==0.1

pillow==12.1.0

numpy==2.4.1
pandas==2.3.3

gunicorn==24.1.1

request==2.31.0